// lib/api/generated/models/BusinessLocation.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BusinessLocation = {
    readonly uuid: string;
    name?: string | null;
    location_data?: any;
    is_active?: boolean;
    readonly created_at: string;
    readonly updated_at: string;
};